Split ``sphinx.ext.intersphinx._shared``
